// src/store/enums/UserActionsEnum.ts
export enum UserActionsTypes {
  GET_ALL_USERS = "GET_ALL_USERS",
  GET_ALL_USERS_SUCCESS = "GET_ALL_USERS_SUCCESS",
  GET_ALL_USERS_FAILURE = "GET_ALL_USERS_FAILURE",
  GET_USER_BY_ID = "GET_USER_BY_ID",
  GET_USER_BY_ID_SUCCESS = "GET_USER_BY_ID_SUCCESS",
  GET_USER_BY_ID_FAILURE = "GET_USER_BY_ID_FAILURE",
  CREATE_USER = "CREATE_USER",
  CREATE_USER_SUCCESS = "CREATE_USER_SUCCESS",
  CREATE_USER_FAILURE = "CREATE_USER_FAILURE",
  UPDATE_USER = "UPDATE_USER",
  UPDATE_USER_SUCCESS = "UPDATE_USER_SUCCESS",
  UPDATE_USER_FAILURE = "UPDATE_USER_FAILURE",
  DELETE_USER = "DELETE_USER",
  DELETE_USER_SUCCESS = "DELETE_USER_SUCCESS",
  DELETE_USER_FAILURE = "DELETE_USER_FAILURE",
  LOGIN_USER = "LOGIN_USER",
  LOGIN_USER_SUCCESS = "LOGIN_USER_SUCCESS",
  LOGIN_USER_FAILURE = "LOGIN_USER_FAILURE",
  CHECK_LOGIN = "CHECK_LOGIN",
  CHECK_LOGIN_SUCCESS = "CHECK_LOGIN_SUCCESS",
  CHECK_LOGIN_FAILURE = "CHECK_LOGIN_FAILURE",
  LOGOUT_USER = "LOGOUT_USER",
  UPDATE_PASSWORD = "UPDATE_PASSWORD",
  UPDATE_PASSWORD_SUCCESS = "UPDATE_PASSWORD_SUCCESS",
  UPDATE_PASSWORD_FAILURE = "UPDATE_PASSWORD_FAILURE",
  FORGOT_PASSWORD = "FORGOT_PASSWORD",
  FORGOT_PASSWORD_SUCCESS = "FORGOT_PASSWORD_SUCCESS",
  FORGOT_PASSWORD_FAILURE = "FORGOT_PASSWORD_FAILURE",
  RESET_PASSWORD = "RESET_PASSWORD",
  RESET_PASSWORD_SUCCESS = "RESET_PASSWORD_SUCCESS",
  RESET_PASSWORD_FAILURE = "RESET_PASSWORD_FAILURE",
}